Refactor tests to extract school models into a separate module
Moved school-related models and fixtures from `conftest.py` to the new `school_models.py` module for better modularity and reuse. Adjusted test imports accordingly to reference the new module location.
