+++ tests/test_all_models.py
@@ -1,8 +1,6 @@
-import pytest
-
 from daomodel import all_models
 from daomodel.db import create_engine, init_db
-from tests.conftest import Person, Book, Hall, Locker, Staff, Student
+from tests.school_models import *
 
 
 @pytest.mark.xfail(reason='Fails unless run individually (to better control loaded models)')
